# backend/src/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "TecnoCel Web API",
    "description": "API REST para la tienda en línea de productos tecnológicos TecnoCel Web. Incluye gestión de productos, clientes, ventas, carrito de compras y más.",
    "version": "1.0.0",
    "contact": {
      "name": "Soporte TecnoCel"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000/api",
      "description": "Servidor de desarrollo"
    }
  ],
  "securityDefinitions": {
    "bearerAuth": {
      "type": "apiKey",
      "name": "Authorization",
      "in": "header",
      "description": "JWT token. Formato: \"Bearer {token}\""
    },
    "bearerAuthCliente": {
      "type": "apiKey",
      "name": "Authorization",
      "in": "header",
      "description": "JWT token para clientes. Formato: \"Bearer {token}\""
    }
  },
  "paths": {
...
    "/envios/admin": {
      "get": {
        "summary": "Listar envíos",
        "description": "Obtiene la lista de envíos con paginación",
        "tags": [
          "Envios"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "query",
            "name": "limite",
            "schema": {
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "schema": {
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "estado",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Lista de envíos obtenida"
          },
          "403": {
            "description": "No tiene permisos"
          },
          "500": {
            "description": "Error interno del servidor"
          }
        }
      }
    },
    "/envios/admin/{id_envio}": {
      "get": {
        "summary": "Obtener detalle de envío",
        "description": "Obtiene los detalles completos de un envío específico",
        "tags": [
          "Envios"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id_envio",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Detalle del envío obtenido"
          },
          "403": {
            "description": "No tiene permisos"
          },
          "404": {
            "description": "Envío no encontrado"
          },
          "500": {
            "description": "Error interno del servidor"
          }
        }
      }
    },
    "/envios/admin/{id_envio}/estado": {
      "patch": {
        "summary": "Actualizar estado de envío",
        "description": "Actualiza el estado de un envío",
        "tags": [
          "Envios"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id_envio",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "nuevoEstado"
                ],
                "properties": {
                  "nuevoEstado": {
                    "type": "string"
                  },
                  "notas": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Estado actualizado exitosamente"
          },
          "400": {
            "description": "Datos inválidos"
          },
          "403": {
            "description": "No tiene permisos"
          },
          "404": {
            "description": "Envío no encontrado"
          },
          "500": {
            "description": "Error interno del servidor"
          }
        }
      }
    },
    "/favoritos/cliente/{id_cliente}": {
      "get": {
        "summary": "Obtener favoritos del cliente",
        "description": "Obtiene la lista de productos favoritos de un cliente",
        "tags": [
          "Favoritos"
        ],
        "security": [
          {
            "bearerAuthCliente": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id_cliente",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Lista de favoritos"
          },
          "500": {
            "description": "Error interno del servidor"
          }
        }
      },
      "post": {
        "summary": "Agregar a favoritos",
        "description": "Agrega un producto a la lista de favoritos del cliente",
        "tags": [
          "Favoritos"
        ],
        "security": [
          {
            "bearerAuthCliente": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id_cliente",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "id_producto"
                ],
                "properties": {
                  "id_producto": {
                    "type": "integer"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Producto agregado a favoritos"
          },
          "400": {
            "description": "Producto ya está en favoritos"
          },
          "404": {
            "description": "Producto no encontrado"
          },
          "500": {
            "description": "Error interno del servidor"
          }
        }
      }
    },
    "/favoritos/cliente/{id_cliente}/estadisticas": {
      "get": {
        "summary": "Obtener estadísticas de favoritos",
        "description": "Obtiene estadísticas sobre los favoritos del cliente",
        "tags": [
          "Favoritos"
        ],
        "security": [
          {
            "bearerAuthCliente": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id_cliente",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Estadísticas de favoritos"
          },
          "500": {
            "description": "Error interno del servidor"
          }
        }
      }
    },
    "/favoritos/cliente/{id_cliente}/producto/{id_producto}": {
      "get": {
        "summary": "Verificar favorito",
        "description": "Verifica si un producto específico está en los favoritos del cliente",
        "tags": [
          "Favoritos"
        ],
        "security": [
          {
            "bearerAuthCliente": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id_cliente",
            "required": true,
            "schema": {
              "type": "integer"
            }
          },
          {
            "in": "path",
            "name": "id_producto",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Estado de favorito verificado"
          },
          "500": {
            "description": "Error interno del servidor"
          }
        }
      },
      "delete": {
        "summary": "Eliminar de favoritos",
        "description": "Elimina un producto de la lista de favoritos del cliente",
        "tags": [
          "Favoritos"
        ],
        "security": [
          {
            "bearerAuthCliente": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id_cliente",
            "required": true,
            "schema": {
              "type": "integer"
            }
          },
          {
            "in": "path",
            "name": "id_producto",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Producto eliminado de favoritos"
          },
          "404": {
            "description": "Favorito no encontrado"
          },
          "500": {
            "description": "Error interno del servidor"
          }
        }
      }
    },
    "/favoritos/cliente/{id_cliente}/producto/{id_producto}/toggle": {
      "put": {
        "summary": "Alternar favorito",
        "description": "Agrega o elimina un producto de favoritos según su estado actual",
        "tags": [
          "Favoritos"
        ],
        "security": [
          {
            "bearerAuthCliente": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id_cliente",
            "required": true,
            "schema": {
              "type": "integer"
            }
          },
          {
            "in": "path",
            "name": "id_producto",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Estado de favorito alternado exitosamente"
          },
          "404": {
            "description": "Producto no encontrado"
          },
          "500": {
            "description": "Error interno del servidor"
          }
        }
      }
    },
...
  },
...
}